+++ lib/ahoy.rb
@@ -2,7 +2,6 @@ require "ipaddr"
 
 require "active_support"
 require "active_support/core_ext"
-require "addressable/uri"
 require "geocoder"
 require "safely/core"